requirements.txt
aiohttp==3.9.1
aiosqlite==0.19.0
discord.py==2.3.2
humanize==4.9.0
python-dotenv==1.0.1
pytz==2023.3
rich==13.7.0
